requirements.txt
beautifulsoup4==4.9.1
PyYAML==5.3.1
requests==2.24.0
slack-webhook==1.0.3
